+++ src/utils.js
@@ -1,0 +1,4 @@
+export const isNumberValue = (value) => {
+  // TODO: 더 좋은 방법? not not 너낌..
+  return !/[^0-9]/g.test(value);
+};
